[#88] Move alternativeFQDN.go to internal/types pkg
