+++ internal/types/alternativeFQDN.go
@@ -1,4 +1,4 @@
-package util
+package types
 
 import "strings"
