+++ internal/types/alternativeFQDN_test.go
@@ -1,4 +1,4 @@
-package util
+package types
 
 import (
 	"reflect"
